requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
celery==4.3.0
django-celery-results==1.1.2
kombu==4.6.6
psycopg2==2.8.4
vine==1.3.0
django-filter==21.1
django-celery-beat==2.0.0
drf-yasg==1.20.0